requirements.txt
fastapi
uvicorn
sqlmodel
jinja2
requests
python-dotenv
